+++ lib/modules/login/utils/roles_pages.dart
@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:ummobile/modules/tabs/models/page_tab_view.dart';
+import 'package:ummobile/modules/tabs/modules/announcements/views/announcements_forms.dart';
 import 'package:ummobile/modules/tabs/modules/calendar/views/calendar_page.dart';
 import 'package:ummobile/modules/tabs/modules/conectate/views/conectate_page.dart';
-import 'package:ummobile/modules/tabs/modules/forms/views/page_forms.dart';
+import 'package:ummobile/modules/tabs/modules/notices/views/notices_page.dart';
+// import 'package:ummobile/modules/tabs/modules/forms/views/page_forms.dart';
 import 'package:ummobile/modules/tabs/modules/payments/views/page_payment.dart';
 import 'package:ummobile/modules/tabs/modules/profile/models/user_credentials.dart';
 import 'package:ummobile/modules/tabs/modules/profile/views/profile_page.dart';
@@ -44,11 +46,21 @@ List<PageTabView> employee = <PageTabView>[
     icon: Icon(Icons.event, size: 30),
     label: "Agenda",
   ),
+  // PageTabView(
+  //   page: FormsPage(),
+  //   icon: Icon(Icons.business_center_rounded, size: 30),
+  //   label: "Formularios",
+  // ),
   PageTabView(
-    page: FormsPage(),
-    icon: Icon(Icons.business_center_rounded, size: 30),
-    label: "Formularios",
+      page: AnnouncementsPage(),
+      icon: Icon(Icons.announcement, size: 30),
+      label: "Anuncios"),
+  PageTabView(
+    page: NoticesPage(),
+    icon: Icon(Icons.article, size: 30),
+    label: "Comunicados",
   ),
+
   PageTabView(
     page: ProfilePage(),
     icon: Icon(AppIcons.user),
