+++ lib/modules/tabs/modules/announcements/views/announcements_forms.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:ummobile/modules/tabs/controllers/navigation_controller.dart';
+
+import 'package:ummobile/modules/app_bar/views/appBar.dart';
+
+class AnnouncementsPage extends StatefulWidget {
+  AnnouncementsPage({Key? key}) : super(key: key);
+
+  @override
+  _AnnouncementsPageState createState() => _AnnouncementsPageState();
+}
+
+class _AnnouncementsPageState extends State<AnnouncementsPage>
+    with AutomaticKeepAliveClientMixin {
+  @override
+  bool get wantKeepAlive => true;
+
+  @override
+  Widget build(BuildContext context) {
+    super.build(context);
+    return Scaffold(
+      appBar: UmAppBar(
+        title: 'de interes'.tr.capitalizeFirst!,
+        leading: IconButton(
+            icon: Icon(Icons.menu),
+            onPressed: () {
+              Get.find<NavigationController>()
+                  .drawerKey
+                  .currentState!
+                  .openDrawer();
+            }),
+      ),
+      body: (const Text('Hola')),
+    );
+  }
+}
