+++ lib/modules/tabs/modules/notices/views/notices_page.dart
@@ -1,0 +1,37 @@
+import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:ummobile/modules/tabs/controllers/navigation_controller.dart';
+
+import 'package:ummobile/modules/app_bar/views/appBar.dart';
+
+class NoticesPage extends StatefulWidget {
+  NoticesPage({Key? key}) : super(key: key);
+
+  @override
+  _NoticesPageState createState() => _NoticesPageState();
+}
+
+class _NoticesPageState extends State<NoticesPage>
+    with AutomaticKeepAliveClientMixin {
+  @override
+  bool get wantKeepAlive => true;
+
+  @override
+  Widget build(BuildContext context) {
+    super.build(context);
+    return Scaffold(
+      appBar: UmAppBar(
+        title: 'anuncios'.tr.capitalizeFirst!,
+        leading: IconButton(
+            icon: Icon(Icons.menu),
+            onPressed: () {
+              Get.find<NavigationController>()
+                  .drawerKey
+                  .currentState!
+                  .openDrawer();
+            }),
+      ),
+      body: (const Text('Hola')),
+    );
+  }
+}
